+++ eras/alonzo/impl/CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add `EraPlutusContext`
 * Add `EraPlutusContext 'PlutusV1` instance to `AlonzoEra`
 * Rename `transTxCert` to `transShelleyTxCert`
+* Remove `witsVKeyNeeded`, in favor of the one from `cardano-ledger-shelley`
 
 ## 1.2.0.0