+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -5,6 +5,8 @@
 * Add `certsTxBodyL` to `EraTxBody`
 * Introduce `TxCert` type family and `EraTxCert` type class.
 * Deprecate `Delegation`
+* Add `toKeyHashWitness`
+* Addition of `getVKeyWitnessTxCert` and `getScriptWitnessTxCert` to `EraTxCert` type class
 
 ## 1.2.0.0